refactor: move feature service to file named features.ts
